+++ CHANGELOG.md
@@ -2,21 +2,48 @@
 
 All notable changes to the "silverfin-development-toolkit" extension will be documented in this file.
 
-## [Unreleased]
+## [1.2.0]
+
+### Updated
+
+- Schema for Liquid Tests updated
+
+## [1.1.1]
+
+### Updated
+
+- Wrong link in README
+
+## [1.1.0]
+
+### Added
+
+- Show only the statusBar icon when there is a YAML file opened in the activeTab
+
+## [1.0.0]
+
+### Added
+
+- SF-Toolkit added as a dependency.
+- Run your Liquid Tests from the Extension.
+- Visualize Liquid Test's output in the YAML files.
 
 ## [0.0.3]
 
 ### Updated
+
 - Replaced path for the JSON Schema (from a local file to an external file). This change will make the schema read-only for users.
 
 ## [0.0.2]
 
 ### Updated
+
 - Local paths for schemas and snippets (`src` directory was omitted when publishing).
 
 ## [0.0.1]
 
 ### Added
+
 - JSON Schema for liquid test written in YAML.
 - YAML Code snippet of a sample test.
 - Red Hat's YAML extension dependency.